+++ R/check.R
@@ -360,10 +360,9 @@ check_max_delay <- function(data,
   obs <- enw_filter_delay(obs, max_delay = daily_max_delay, timestep = "day")
 
   # filter by earliest observed report date
-  obs <- obs[,
-    .SD[reference_date >= min(report_date) | is.na(reference_date)],
-    by = .group
-  ]
+  obs <- enw_filter_reference_dates_by_report_start(
+    obs, by = ".group", copy = FALSE
+  )
 
   latest_obs <- enw_latest_data(obs)
   fully_observed_date <- latest_obs[, max(report_date)] - max_delay_obs_q + 1
@@ -598,7 +597,7 @@ check_numeric_timestep <- function(dates, date_var, timestep, exact = TRUE) {
       "{date_var} does not have the specified timestep of {timestep} day(s)"
     )
   }
-  return(invisible(NULL))
+  invisible(NULL)
 }
 
 #' Check timestep
@@ -638,17 +637,15 @@ check_timestep <- function(obs, date_var, timestep = "day", exact = TRUE,
   dates <- sort(dates)
   dates <- dates[!is.na(dates)]
 
-  if (length(dates) <= 1 && !check_nrow) {
-    return(invisible(NULL))
-  }
   if (length(dates) <= 1) {
-    cli::cli_abort("There must be at least two observations")
+    if (check_nrow) {
+      cli::cli_abort("There must be at least two observations")
+    }
+  } else {
+    internal_timestep <- get_internal_timestep(timestep)
+    check_numeric_timestep(dates, date_var, internal_timestep, exact)
   }
-
-  internal_timestep <- get_internal_timestep(timestep)
-  check_numeric_timestep(dates, date_var, internal_timestep, exact)
-
-  return(invisible(NULL))
+  invisible(NULL)
 }
 
 #' Check timestep by group
@@ -678,7 +675,7 @@ check_timestep_by_group <- function(obs, date_var, timestep = "day",
     by = ".group"
   ]
 
-  return(invisible(NULL))
+  invisible(NULL)
 }
 
 #' Check timestep by date
@@ -721,7 +718,7 @@ check_timestep_by_date <- function(obs, timestep = "day", exact = TRUE) {
     ),
     by = c("report_date", ".group")
   ]
-  return(invisible(NULL))
+  invisible(NULL)
 }
 
 #' Check observation indicator
@@ -746,7 +743,7 @@ check_observation_indicator <- function(
     !is.logical(new_confirm[[observation_indicator]])) {
     cli::cli_abort("observation_indicator must be a logical")
   }
-  return(invisible(NULL))
+  invisible(NULL)
 }
 
 #' Check design matrix sparsity
@@ -770,21 +767,20 @@ check_observation_indicator <- function(
 check_design_matrix_sparsity <- function(matrix, sparsity_threshold = 0.9,
                                          min_matrix_size = 50,
                                          name = "checked") {
-  if (length(matrix) < min_matrix_size) {
-    return(invisible(NULL))
-  }
+  if (length(matrix) >= min_matrix_size) {
+    zero_proportion <- sum(matrix == 0) / length(matrix)
 
-  zero_proportion <- sum(matrix == 0) / length(matrix)
-
-  if (zero_proportion > sparsity_threshold) {
-    cli::cli_alert_info(
-      c(
-        "The {name} design matrix is sparse (>{sparsity_threshold*100}% ",
-        "zeros). Consider using `sparse_design = TRUE` in `enw_fit_opts()` ",
-        "to potentially reduce memory usage and computation time."
+    if (zero_proportion > sparsity_threshold) {
+      cli::cli_alert_info(
+        c(
+          "The {name} design matrix is sparse ",
+          "(>{sparsity_threshold*100}% zeros). Consider ",
+          "using `sparse_design = TRUE` in ",
+          "`enw_fit_opts()` to potentially reduce memory ",
+          "usage and computation time."
+        )
       )
-    )
+    }
   }
-
-  return(invisible(NULL))
+  invisible(NULL)
 }