+++ tests/testthat/test-enw_incidence_to_linelist.R
@@ -1,5 +1,8 @@
 test_that("enw_incidence_to_linelist can return a linelist", {
-  incidence <- enw_add_incidence(germany_covid19_hosp)
+  incidence <- enw_filter_reference_dates_by_report_start(
+    germany_covid19_hosp
+  )
+  incidence <- enw_add_incidence(incidence)
   incidence <- enw_filter_reference_dates(
     incidence[location == "DE"],
     include_days = 10